requirements.txt
tensorflow
scipy
Pillow
opencv-python